college-attendance-api/requirements.txt
Flask==3.1.2
flask-cors==5.0.0
gunicorn==20.1.0
pytz==2024.1
PyJWT==2.8.0
requests==2.31.0
cryptography==42.0.5
langchain-google-genai==2.1.2
langchain-core==0.3.51
langgraph==0.2.56